requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.2
Automat==20.2.0
autopep8==1.4.4
backcall==0.1.0
boto3==1.12.11
botocore==1.15.11
certifi==2019.11.28
cffi==1.14.0
channels==2.2.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
decorator==4.4.2
dj-database-url==0.5.0
Django==2.2.9
django-cors-headers==3.2.1
django-localflavor==3.0
django-multiselectfield==0.1.12
django-storages==1.9.1
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
docxtpl==0.6.3
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
hiredis==1.0.1
httplib2==0.17.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.3.1
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-docx==0.8.7
python-magic==0.4.15
python-stdnum==1.13
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sentry-sdk==0.14.2
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
Twisted==19.10.0
txaio==20.1.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
whitenoise==4.0
xlrd==1.2.0
zope.interface==4.7.1
